fix(providers): bound HTTP response-body reads to prevent OOM
Every integration provider except WeChat read error-response bodies with the
unbounded `response.text().await` (or `.bytes()`) before truncating the
message — pulling the ENTIRE body into memory first. A hostile or compromised
target endpoint could OOM the gateway by replying with a giant body. For the
webhook provider the endpoint URL is caller-supplied, so this is a real
cross-tenant DoS, not just a malicious-upstream concern.

Add a shared `read_bounded_body(response, max_bytes)` to `acteon-provider`
(gated on the `reqwest` feature every provider already pulls via
`trace-context`) — it pumps `Response::chunk()` and stops at the cap instead
of reading the whole body — plus `MAX_ERROR_BODY_READ_BYTES` (2 KiB) and
`MAX_RESPONSE_BODY_READ_BYTES` (1 MiB) consts. WeChat's local copy is removed
in favor of the shared one.

- All 10 unbounded providers now read error bodies via the bounded helper.
- The webhook + Teams SUCCESS bodies (returned to the caller) are also bounded
  at 1 MiB — webhook because its endpoint is caller-controlled, Teams because
  its success body is a trivial `"1"`. Other providers' success paths parse a
  small JSON status with `response.json()` from trusted vendor APIs and are
  left unchanged (matching the established WeChat pattern).

Follow-up: S3 `get_object` / GCS `download_object` stream whole objects into a
`Vec<u8>` (a different fix — size cap or streaming).
