+++ crates/provider/src/lib.rs
@@ -9,8 +9,14 @@ pub mod resource_lookup;
 #[cfg(feature = "webhook")]
 pub mod webhook;
 
+// Shared HTTP response helpers (bounded body reads) — require reqwest.
+#[cfg(feature = "reqwest")]
+pub mod http;
+
 pub use context::DispatchContext;
 pub use error::{ProviderError, truncate_error_body};
+#[cfg(feature = "reqwest")]
+pub use http::{MAX_ERROR_BODY_READ_BYTES, MAX_RESPONSE_BODY_READ_BYTES, read_bounded_body};
 pub use log::LogProvider;
 pub use provider::{DynProvider, Provider};
 pub use registry::ProviderRegistry;
